reducer 분리
