+++ src/containers/CounterContainer.js
@@ -26,8 +26,8 @@ export function getRandomColor() {
 
 //store 안의 state 값을 props 로 연결
 const mapStateToProps = (state) => ({
-    color: state.color,
-    number: state.number
+    color: state.colorData.color,
+    number: state.numberData.number
 });
 
 //액션 생성함수를 사용하여 액션을 생성하고,
