+++ src/reducers/color.js
@@ -1,0 +1,18 @@
+import * as types from '../actions/ActionType';
+
+const initialState = {
+    color: 'black'
+};
+
+const color = (state = initialState, action) => {
+    switch(action.type) {
+        case types.SET_COLOR:
+            return {
+                color: action.color
+            };
+        default:
+            return state;
+    }
+};
+
+export default color;
