+++ src/reducers/index.js
@@ -1,30 +1,17 @@
-import * as types from '../actions/ActionType';
+import color from './color';
+import number from './number';
 
-const initialState = {
-    number: 0,
-    color: 'black'
-};
+import { combineReducers } from 'redux';
 
-function counter(state = initialState, action) {
-    switch (action.type) {
-        case types.INCREMENT:
-            return {
-                ...state,
-                number: state.number + 1
-            };
-        case types.DECREMENT:
-            return {
-                ...state,
-                number: state.number - 1
-            };
-        case types.SET_COLOR:
-            return {
-                ...state,
-                color: action.color
-            };
-        default:
-            return state;
-    }
-};
+// const initialState = {
+//     number: 0,
+//     color: 'black'
+// };
 
-export default counter;
+const reducers = combineReducers({
+    numberData: number,
+    colorData: color
+});
+
+
+export default reducers;
