+++ src/reducers/number.js
@@ -1,0 +1,24 @@
+import * as types from '../actions/ActionType';
+
+const initialState = {
+    number: 0
+};
+
+const number = (state = initialState, action) => {
+    switch (action.type) {
+        case types.INCREMENT:
+            return {
+                ...state,
+                number: state.number + 1
+            };
+        case types.DECREMENT:
+            return {
+                ...state,
+                number: state.number - 1
+            };
+        default:
+            return state;
+    }
+};
+
+export default number;
